Added method to get students of all classes on stream
