+++ src/app/services/student.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 import { ClassService } from './class.service';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, tap } from 'rxjs/operators';
 import { ClassData } from '../models/class-data';
+import { StudentsOfStream } from '../models/students-of-stream';
 
 @Injectable({
   providedIn: 'root'
@@ -12,15 +13,41 @@ export class StudentService {
   constructor(private classService: ClassService) { }
 
   /**
-   * Method returns active students
-   * @returns - Number of active classes
+   * Method returns students
+   * @returns - Number of students
    */
-  public getStudents(type: string): Observable<number> {
+  public getNumberOfStudents(type: string): Observable<number> {
     return this.classService.getClasses(type)
       .pipe(
         map((result: ClassData[]) => {
           return result.reduce((students, currClass) => students + currClass.numOfStudents, 0);
         })
+      );
+  }
+  public getStudentsByStream(stream: number): any {
+    return this.classService.getClasses('active')
+      .pipe(
+        map(result => {
+          const uniqueClasses: ClassData[] = [];
+          result
+            .filter(item => item.isActive && parseInt(item.className, 10) === stream && item.numOfStudents)
+            .forEach(item => {
+              if (!uniqueClasses.some(value => item.className === value.className)) {
+                uniqueClasses.push(item);
+              }
+            });
+          const studentsStream: StudentsOfStream[] = uniqueClasses.map(item => {
+            return {
+              className: item.className,
+              numOfStudents: item.numOfStudents
+            };
+          });
+          return {
+            studentsData: studentsStream,
+            allStudents: studentsStream.reduce((sum, studentClass) => sum + studentClass.numOfStudents , 0)
+          };
+        }),
+        tap(result => console.log(result))
       );
   }
 }
